- Basic with annotate function structure - PySpark Basic model port
