+++ python/run-tests.py
@@ -2,7 +2,6 @@ from test.annotators import *
 from test.misc import *
 
 # Annotator tests
-
 unittest.TextTestRunner().run(BasicAnnotatorsTestSpec())
 unittest.TextTestRunner().run(RegexMatcherTestSpec())
 unittest.TextTestRunner().run(LemmatizerTestSpec())
@@ -19,3 +18,4 @@ unittest.TextTestRunner().run(ParamsGettersTestSpec())
 unittest.TextTestRunner().run(UtilitiesTestSpec())
 unittest.TextTestRunner().run(ConfigPathTestSpec())
 unittest.TextTestRunner().run(SerializersTestSpec())
+unittest.TextTestRunner().run(ModelTestSpec())
