+++ python/sparknlp/internal.py
@@ -56,13 +56,13 @@ class _DownloadModel(ExtendedJavaWrapper):
 
 class _DownloadPipeline(ExtendedJavaWrapper):
     def __init__(self, name, language):
-        super(_DownloadPipeline, self).__init__("com.johnsnowlabs.pretrained.PythonResourceDownloader.downloadPipeline")
+        super(_DownloadPipeline, self).__init__("com.johnsnowlabs.downloader.PythonResourceDownloader.downloadPipeline")
         self._java_obj = self._new_java_obj(self._java_obj, name, language)
 
 
 class _ClearCache(ExtendedJavaWrapper):
     def __init__(self, name, language):
-        super(_ClearCache, self).__init__("com.johnsnowlabs.pretrained.PythonResourceDownloader.clearCache")
+        super(_ClearCache, self).__init__("com.johnsnowlabs.downloader.PythonResourceDownloader.clearCache")
         self._java_obj = self._new_java_obj(self._java_obj, name, language)
